modify local settings for python3

+++ train/train.py
@@ -1,5 +1,4 @@
 ''' Training Frustum PointNets.
-
 Author: Charles R. Qi
 Date: September 2017
 '''
@@ -216,7 +215,7 @@ def train_one_epoch(sess, ops, train_writer):
     # Shuffle train samples
     train_idxs = np.arange(0, len(TRAIN_DATASET))
     np.random.shuffle(train_idxs)
-    num_batches = len(TRAIN_DATASET)/BATCH_SIZE
+    num_batches = int(len(TRAIN_DATASET)/BATCH_SIZE)
 
     # To collect statistics
     total_correct = 0
